some basic functionalities are working properly now

+++ memory_management/vaddr_tracereader.c
@@ -75,7 +75,7 @@ int NextAddress(FILE *trace_file, p2AddrTr *addr_ptr) {
  */
 void AddressDecoder(p2AddrTr *addr_ptr, FILE *out) {
   
-  fprintf(out, "%08lx ", addr_ptr->addr);	/* address */
+  fprintf(out, "%08lx ", (long unsigned int) addr_ptr->addr);	/* address */
   /* what type of address request */
   switch (addr_ptr->reqtype) {
     case FETCH:
@@ -137,7 +137,7 @@ void AddressDecoder(p2AddrTr *addr_ptr, FILE *out) {
      timestamp
   */
   fprintf(out, "%2d\t%02x\t%1d\t%08lx\n", addr_ptr->size, addr_ptr->attr,
-	  addr_ptr->proc, addr_ptr->time);
+	  addr_ptr->proc,(long unsigned int) addr_ptr->time);
 }
